requirements.txt
altgraph==0.17.3
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
Flask==2.2.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pefile==2023.2.7
pycparser==2.21
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyMySQL==1.0.2
python-dotenv==1.0.0
pywin32-ctypes==0.2.0
Werkzeug==2.2.3
zipp==3.15.0
